feat: Upgrade to Spring Boot 3.5.14 (#99)
* feat: Upgrade to Java 21, Spring Boot 3.3.11;

* feat: change java version to 21 in maven-release-v2.yaml, dev-mvn-docker-build.yaml;

* fix(deps): Exclude dependency org.antlr.antlr4-runtime,

* fix(deps): Bump APK sops to 3.11.0-r4, libpng to 1.6.55-r0, libexpat to 2.7.4-r0, zlib to 1.3.2-r0. Base JDK image is changed to bellsoft/liberica-openjdk-alpine-musl:21.0.10.

* fix(security): Bump dependencies: jackson to 2.18.6, undertow to 2.3.24.Final.

* fix(security): Dockerfile: Bump APK package libexpat to 2.7.5-r0.

* fix(security): Bump dependencies: tomcat-embed to 10.1.54, io.netty to 4.1.132.Final, guava to 32.0.1-jre.

* fix(security): Bump dependencies: com.hazelcast:hazelcast to 5.2.5, snakeyaml to 2.0, org.lz4:lz4-java to 1.8.1.

* fix: SPRING_PROFILES environment property is replaced to ACTIVE_PROFILES_SPRING.

* docs: MARKDOWN and NATURAL_LANGUAGE errors are fixed.

* fix: httpclient dependency is removed, httpclient5 is added instead. Also: TreeNodeUtilsTest: code is cleansed.

* fix: Fix activemq connection issue

* feat: Upgrade to Spring Boot 3.5.11: Bump spring-boot, spring-cloud and related dependencies versions. Improve dependency management. Migrate tests from Junit to Jupiter. Remove obsolete DisableAutoGeneratedTagsPlugin.java. Migrate from Springfox v2 to Springdoc v3. Add exclusions of transitive dependencies to junit, commons-logging, javax.annotations etc.

* fix(security): Bump APK packages: libpng to 1.6.57-r0, libcrypto3 to 3.5.6-r0, libssl3 to 3.5.6-r0, sops to 3.11.0-r5, musl to 1.2.5-r23, musl-utils to 1.2.5-r23.

* fix(security): Bump netty to 4.1.132.Final, httpclient5/httpcore5 to 5.4.4/5.3.4, kafka-clients to 3.9.2.

* fix(security): Bump Spring Boot to 3.5.12, Spring Cloud to 2025.0.2, activemq to 6.2.4, tomcat to 10.1.54, gson to 2.13.2. Also: au.com.dius.pact.provider:junit5spring and :junit5 to 4.6.15, mockito-junit-jupiter to 5.23.0, reactor-netty to 1.2.17, and some unused old dependencies are removed.

* fix(tests): Pact test (provider side) is fixed after au.com.dius.pact.provider dependencies upgrade.

* fix(security): Bump openjdk to 21.0.11.

* feat: change version spring-boot to 3.5.14

* feat: change APK versions curl to 8.19.0-r0; libexpat to 2.8.1-r0 , libpng to 1.6.58-r1; sops to 3.11.0-r6; xz-libs to 5.8.3-r0;

* feat: add properties 'spring.cloud.openfeign.httpclient.disable-ssl-validation', 'spring.cloud.openfeign.httpclient.hc5.enabled', 'spring.cloud.openfeign.okhttp.enabled' to application.properties;

* feat: fix cast long to biginteger for report context,

* feat: change APK curl to 8.20.0-r0, libexpat to 2.8.2-r0;

* feat: fix warning in logback config;

* fix(deps): Bump itf-core to 5.0.0, Undertow to 2.3.24.Final, lombok to 1.18.46, httpclient5/httpcore5 to 5.5.2/5.3.6, spring-boot-maven-plugin to 3.5.14. Remove unused outdated cglib:cglib:3.1 and spring-cloud-starter-bootstrap dependencies. Make spring-kafka Spring-Boot-managed.

* fix: Remove unused migration script.

---------

+++ reports-aggregator/src/main/java/org/qubership/automation/itf/ui/tree/TreeNodeUtils.java
@@ -22,7 +22,7 @@ import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 
-import javax.annotation.Nonnull;
+import jakarta.annotation.Nonnull;
 
 import org.qubership.automation.itf.report.model.ReportObject;
 
@@ -69,8 +69,8 @@ public class TreeNodeUtils {
         }
         //TODO It's hardcode. Need Use @SqlResultSetMapping in ReportObject
         ReportObject reportObject = new ReportObject();
-        reportObject.setId((BigInteger) object[0]);
-        reportObject.setParent((BigInteger) object[1]);
+        reportObject.setId(BigInteger.valueOf(((Long) object[0])));
+        reportObject.setParent(BigInteger.valueOf(((Long) object[1])));
         reportObject.setType((String) object[2]);
         reportObject.setDescription((String) object[3]);
         reportObject.setPath((String) object[4]);
